moved FormattingWorker to correct namespace
